IDokanOperations2 library implementation
* IDokanOperations2 is a new way to implement Dokan file systems in .NET with less CPU and memory pressure.
* Includes several other minor optimizations as well.
* Existing IDokanOperation and IDokanOperationUnsafe implementations get wrapped in a compatibility layer called DokanOperationsAdapter, that in turn implements the new IDokanOperations2 interface.
* Removed "-windows" suffix from target frameworks to allow library to be consumed by applicaitons that chose between Windows and other OS implementations at runtime.
* Removed .NET Framework 4.0 from target frameworks and added 4.6 and 4.8 (last supported on Windows Vista and last supported on Windows 7).
* DokanNetMirror with IDokanOperations2. Old DokanNetMirror moved to DokanNetMirrorLegacy.
* Added some logic to keep a reference to DokanInstance and associated operations alive while file system is mounted, to avoid crashes in native code if the last reference is collected by GC.

+++ DokanNet.Tests/BufferPoolTests.cs
@@ -42,7 +42,7 @@ namespace DokanNet.Tests
             for (int i = 0; i < 1000; i++)
             {
                 int size = random.Next(0, 2 * MB);
-                buffer = pool.RentBuffer((uint)size, logger);
+                buffer = pool.RentBuffer(size, logger);
                 Assert.AreEqual(size, buffer.Length, "Wrong buffer size.");
                 pool.ReturnBuffer(buffer, logger);
             }
